requirements.txt
streamlit==1.37.1
requests==2.32.3
fastapi==0.112.2
uvicorn==0.30.6
openai==0.28
google-generativeai==0.7.2
python-dotenv==1.0.1